+++ org.palladiosimulator.analyzer.slingshot.planner/src/org/palladiosimulator/analyzer/slingshot/planner/data/HDDContainerSpecification.java
@@ -4,7 +4,7 @@ public class HDDContainerSpecification extends ContainerSpecification {
 	private double writeProcessingRate;
 	private double readProcessingRate;
 
-	public HDDContainerSpecification(String id, int replicas, int processingRate, String schedulingPolicy, double MTTR,
+	public HDDContainerSpecification(String id, int replicas, double processingRate, String schedulingPolicy, double MTTR,
 			double MTTF, double writeProcessingRate, double readProcessingRate) {
 		super(id, replicas, processingRate, schedulingPolicy, MTTR, MTTF);
 		this.writeProcessingRate = writeProcessingRate;
